% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compBeta}
\alias{compBeta}
\title{Compute log odds ratio}
\usage{
compBeta(theta)
}
\arguments{
\item{theta}{is the vector with multinomial counterfactual distribution, P(Y <= k)}
}
\value{
point estimate
}
\description{
Compute log odds ratio
}
