% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimaBeta}
\alias{estimaBeta}
\title{Enhanded efficiency TMLE of the Log-odds ratio}
\usage{
estimaBeta(data, Qkn1, Qkn0, gDn1, gDn0, gAn1)
}
\arguments{
\item{data}{is a data frame with n rows with variables A (treatment variable), D (missingness indicator), and Y (outcome)}

\item{Qkn1}{n x (k-1) matrix containing P(Y=k|A=1,Y>=k,W)}

\item{Qkn0}{Qkn0 same as above for A=0}

\item{gDn1}{is a vector of sicze n containing Pr(D=1 | A=1, W)}

\item{gDn0}{is a vector containing Pr(D=1 | A=0, W)}

\item{gAn1}{is a ectoe containing the propensity scores Pr(A=1|W)}
}
\value{
point estimates and standard error
}
\description{
Enhanded efficiency TMLE of the Log-odds ratio
}
