% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{iptw}
\alias{iptw}
\title{Inverse Probability Weighted Estimator}
\usage{
iptw(data, gDn1, gDn0, gAn1)
}
\arguments{
\item{data}{is a data frame with n rows with variables A (treatment variable), D (missingness indicator), and Y (outcome)}

\item{gDn1}{is a vector of sicze n containing Pr(D=1 | A=1, W)}

\item{gDn0}{is a vector containing Pr(D=1 | A=0, W)}

\item{gAn1}{is a ectoe containing the propensity scores Pr(A=1|W)}
}
\value{
point estimate
}
\description{
Inverse Probability Weighted Estimator
}
